src/requirements.txt
django>=2.1.6
pytz==2018.9
social-auth-app-django==3.1.0
djangorestframework==3.9.1
gunicorn==19.9.0
cloudstorage==0.9.0
google-cloud-storage==1.14.0
python-magic-bin==0.4.14
pillow==6.0.0